Add a new functionality to fast zoom to a selected label.
When a label is selected, the key combination Alt+C brings the corresponding rectangle into zoomed focus
The zoom level can be controlled by the config key `fast_zoom_target_size`.
The valid values for the config key are floats between 0 and 1. It controls what percentage of the view will be occupied by the target rectangle after zooming.

To achieve this, added a keyPressEvent t the app that calls on a geometry mapping from the canvas.